requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
motor==3.3.2
pymongo==4.6.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Environment Management
python-dotenv==1.0.1

# AI/ML
groq==0.4.2

# Utilities
python-dateutil==2.8.2
